+++ R/incidence.R
@@ -14,7 +14,7 @@
 #' @param interval An integer or character indicating the (fixed) size of the
 #' time interval used for computing the incidence; defaults to 1 day. This can
 #' also be a text string that corresponds to a valid date interval: day, week,
-#' month, quarter, or year. (See Note)
+#' month, quarter, or year. (See Note).
 #'
 #' @param groups An optional factor defining groups of observations for which
 #' incidence should be computed separately.
@@ -34,21 +34,29 @@
 #'
 #'
 #' - **dates**: The dates marking the left side of the bins used for counting
-#' events. When ISO week-based weekly incidence is computed, the dates are the
-#' first days of corresponding isoweeks.
-#'
+#' events. When `standard = TRUE` and the interval represents weeks, months,
+#' quarters, or years, the first date will represent the first standard date
+#' (See Interval specification, below).
+#'    
 #' - **counts**: A matrix of incidence counts, which one column per group (and
 #' a single column if no groups were used).
 #'
 #' - **timespan**: The length of the period for which incidence is computed, in
 #' days.
 #'
-#' - **interval**: The bin size, in number of days; e.g. 7 indicates weekly
-#' incidence.
+#' - **interval**: The bin size. If it's an integer, it represents the number
+#'   of days between each bin. It can also be a character, e.g. "2 weeks" or 
+#'   "6 months".
 #'
 #' - **n**: The total number of cases.
 #'
-#' - **isoweeks**: ISO 8601 week format yyyy-Www, which is returned only when
+#' - **weeks**: Dates in week format (YYYY-Www), where YYYY corresponds to the
+#'   year of the given week and ww represents the numeric week of the year. 
+#'   This will be a produced from the function [aweek::date2week()]. Note that
+#'   these will have a special `"week_start"` attribute indicating which day of
+#'   the ISO week the week starts on (see Weeks, below).
+#'   
+#' - **isoweeks**: ISO 8601 week format YYYY-Www, which is returned only when
 #' ISO week-based weekly incidence is computed.
 #'
 #'
@@ -62,18 +70,53 @@
 #' }
 #' 
 #' \subsection{Interval specification (`interval`)}{
-#' If `interval` is a valid character (e.g. "week" or "month"), then
+#' If `interval` is a valid character (e.g. "week" or "1 month"), then
 #' the bin will start at the beginning of the interval just before the first
 #' observation by default. For example, if the first case was recorded on
 #' Wednesday, 2018-05-09:
 #'
-#'  - "week"    : first day of the ISOweek (i.e. Monday, 2018-05-07)
+#'  - "week"    : first day of the week (i.e. Monday, 2018-05-07) (defaults to ISO weeks, see "Week intervals", below)
 #'  - "month"   : first day of the month (i.e. 2018-05-01)
 #'  - "quarter" : first day of the quarter (i.e. 2018-04-01)
 #'  - "year"    : first day of the calendar year (i.e. 2018-01-01)
 #'
 #' These default intervals can be overridden with `standard = FALSE`, which
 #' sets the interval to begin at the first observed case.
+#' }
+#' 
+#' \subsection{Week intervals}{
+#'
+#' As of _incidence_ version 1.7.0, it is possible to construct standardized
+#' incidence objects standardized to any day of the week thanks to the
+#' [aweek::date2week()] function from the \pkg{aweek} package. The default
+#' state is to use ISO 8601 definition of weeks, which start on Monday. You can
+#' specify the day of the week an incidence object should be standardised to by
+#' using the pattern "{n} {W} weeks" where "{W}" represents the weekday in an
+#' English or current locale and "{n}" represents the duration, but this can be
+#' ommitted.  Below are examples of specifying weeks starting on different days
+#' assuming we had data that started on 2016-09-05, which is ISO week 36 of
+#' 2016:
+#'
+#'  - interval = "2 monday weeks" (Monday 2016-09-05)
+#'  - interval = "1 tue week" (Tuesday 2016-08-30)
+#'  - interval = "1 Wed week" (Wednesday 2016-08-31)
+#'  - interval = "1 Thursday week" (Thursday 2016-09-01) 
+#'  - interval = "1 F week" (Friday 2016-09-02)
+#'  - interval = "1 Saturday week" (Saturday 2016-09-03)
+#'  - interval = "Sunday week" (Sunday 2016-09-04)
+#'
+#' It's also possible to use something like "3 weeks: Saturday"; In addition,
+#' there are keywords reserved for specific days of the week:
+#'   
+#'   - interval = "week", standard = TRUE (Default, Monday)
+#'   - interval = "ISOweek"  (Monday)
+#'   - interval = "EPIweek"  (Sunday)
+#'   - interval = "MMWRweek" (Sunday)
+#'
+#' The "EPIweek" specification is not strictly reserved for CDC epiweeks, but
+#' can be prefixed (or posfixed) by a day of the week: "1 epiweek: Saturday".
+#'
+#' }
 #'
 #' \subsection{The `first_date` argument}{
 #' Previous versions of _incidence_ had the `first_date` argument override
@@ -88,7 +131,7 @@
 #' number of days they encompass and warnings will be generated when the first
 #' date falls outside of a calendar date that is easily represented across the
 #' interval.
-#' }
+#' 
 #'
 #' @seealso
 #' The main other functions of the package include:
@@ -147,18 +190,31 @@
 #'   inc.week
 #'   plot(inc.week)
 #'   plot(inc.week, border = "white") # with visible border
-#'   inc.isoweek <- incidence(onset, interval = 7, standard = TRUE)
+#'   
+#'   # Starting on Monday
+#'   inc.isoweek <- incidence(onset, interval = "isoweek")
 #'   inc.isoweek
+#'   
+#'   # Starting on Sunday
+#'   inc.epiweek <- incidence(onset, interval = "epiweek")
+#'   inc.epiweek
+#' 
+#'   # Starting on Saturday
+#'   inc.epiweek <- incidence(onset, interval = "saturday epiweek")
+#'   inc.epiweek
+#'
 #'   ## use group information
-#'   sex <- ebola_sim$linelist$gender
+#'   sex <- outbreaks::ebola_sim$linelist$gender
 #'   inc.week.gender <- incidence(onset, interval = 7,
 #'                                groups = sex, standard = FALSE)
 #'   inc.week.gender
 #'   head(inc.week.gender$counts)
-#'   plot(inc.week.gender)
-#'   inc.isoweek.gender <- incidence(onset, interval = 7,
-#'                                   groups = sex, standard = TRUE)
-#'   inc.isoweek.gender
+#'   plot(inc.week.gender, border = "grey90")
+#'   inc.satweek.gender <- incidence(onset, interval = "2 epiweeks: saturday",
+#'                                   groups = sex)
+#'   inc.satweek.gender
+#'   plot(inc.satweek.gender, border = "grey90")
+#'
 #' })}
 #' 
 #' # Use of first_date
@@ -202,10 +258,11 @@ incidence.Date <- function(dates, interval = 1L, standard = TRUE, groups = NULL,
     fd  <- as.character(deparse(the_call[["first_date"]]))
     msg <- "\n\nAs of incidence version 1.6.0, the default behavior has been"
     msg <- paste(msg, "modified so that `first_date` no longer overrides")
-    msg <- paste(msg, "`standard`. If you want to use %s as the precise")
+    msg <- paste(msg, "`standard`. This means that the first date will be")
+    msg <- paste(msg, "either on or before %s.\nIf you want to use %s as the precise")
     msg <- paste(msg, "`first_date`, set `standard = FALSE`.")
     msg <- paste(msg, "To remove this warning in the future,  explicitly set the `standard` argument OR use `options(incidence.warn.first_date = FALSE)`\n", sep = "\n\n")
-    warning(sprintf(msg, fd))
+    warning(sprintf(msg, first_date, fd))
     # turn the warning off so that it's not so noisy
     options(incidence.warn.first_date = FALSE)
   }
@@ -228,7 +285,11 @@ incidence.Date <- function(dates, interval = 1L, standard = TRUE, groups = NULL,
                         ...)
   if (check_week(interval) && standard) {
     # dates are the first days of corresponding ISOweeks.
-    out$isoweeks <- substr(ISOweek::date2ISOweek(out$dates), 1, 8)
+    week_start   <- get_week_start(interval)
+    out$weeks <- aweek::date2week(out$dates, week_start, floor_day = TRUE)
+    if (attr(out$weeks, "week_start") == 1) {
+      out$isoweeks <- as.character(out$weeks)
+    }
   }
 
   out
@@ -316,7 +377,7 @@ incidence.POSIXt <- function(dates, interval = 1L, standard = TRUE, groups = NUL
   ## make sure input can be used
 
   dots  <- check_dots(list(...), names(formals(incidence.Date)))
-  dates <- check_dates(dates)
+  dates <- check_dates(as.POSIXct(dates))
 
   ret <- incidence(as.Date(dates),
                    interval = interval,