+++ R/make_breaks.R
@@ -23,7 +23,10 @@ make_breaks_easier <- function(dates, the_interval, first_date = NULL,
 
   the_interval    <- valid_interval_character(the_interval)
   date_interval   <- is.character(the_interval) && is_date_interval(the_interval)
-  uneven_interval <- date_interval && the_interval %in% c("month", "quarter", "year")
+  is_month        <- grepl("month", the_interval, ignore.case = TRUE)
+  is_quarter      <- grepl("quarter", the_interval, ignore.case = TRUE)
+  is_year         <- grepl("year", the_interval, ignore.case = TRUE)
+  uneven_interval <- date_interval && (is_month || is_quarter || is_year)
 
   # getting information about the date
   fd        <- as.character(first_date)
@@ -34,23 +37,22 @@ make_breaks_easier <- function(dates, the_interval, first_date = NULL,
     if (isTRUE(dots$standard)) {
       is_a_week <- !uneven_interval && check_week(the_interval)
       if (is_a_week) {
-        # This returns something like 2018-W29-2, where the last digit indicates
-        # the day of the week
-        first_isoweek <- ISOweek::date2ISOweek(first_date)
-        # Here, we force it to be the start of the week
-        substr(first_isoweek, 10, 10) <- "1"
-        # and convert it back
-        first_date <- ISOweek::ISOweek2date(first_isoweek)
+        week_start    <- get_week_start(the_interval)
+        the_interval  <- get_week_duration(the_interval)
+        # This returns something like 2018-W29
+        first_isoweek <- aweek::date2week(first_date, week_start, floor_day = TRUE)
+        # here we convert it back to a date
+        first_date    <- aweek::week2date(first_isoweek)
       }
       if (uneven_interval) {
         # Replace the day with the first day of the month
         substr(fd, 9, 10) <- "01"
-        if (the_interval == "quarter") {
+        if (is_quarter) {
           # Replace the month with the first month of the quarter
           m <- (as.integer(substr(fd, 6, 7)) - 1L) %/% 3L
           substr(fd, 6, 7) <- sprintf("%02d", (m * 3) + 1L)
         }
-        if (the_interval == "year") {
+        if (is_year) {
           # Replace the month with the first month of the year
           substr(fd, 6, 7) <- "01"
         }
@@ -58,7 +60,7 @@ make_breaks_easier <- function(dates, the_interval, first_date = NULL,
         first_date <- as.Date(fd)
       }
     } else {
-      if (uneven_interval && the_interval != "year" && the_day > 28) {
+      if (uneven_interval && !is_year && the_day > 28) {
         # The first date represents a day that doesn't occur in all months
         msg <- paste("The first_date (%s) represents a day that does not",
                      "occur in all months. Because of this, bins may not",
@@ -69,7 +71,7 @@ make_breaks_easier <- function(dates, the_interval, first_date = NULL,
         msg <- paste(strwrap(msg), collapse = "\n")
         warning(sprintf(msg, fd), call. = FALSE)
       }
-      if (the_interval == "year" && the_month == 2 && the_day == 29) {
+      if (is_year && the_month == 2 && the_day == 29) {
         # The first date occurs on a leap day.
         msg <- paste("The first_date (%s) represents a day that does not",
                      "occur in all years. Because of this, bins may not",