+++ R/print.R
@@ -5,9 +5,10 @@ print.incidence <- function(x, ...) {
   cat("<incidence object>\n")
   cat(sprintf("[%d cases from days %s to %s]\n",
               sum(x$n), min(x$dates), max(x$dates)))
-  if ("isoweeks" %in% names(x)) {
-    cat(sprintf("[%d cases from ISO weeks %s to %s]\n",
-                sum(x$n), head(x$isoweeks, 1), tail(x$isoweeks, 1)))
+  if ("weeks" %in% names(x)) {
+    type_of_week <- get_type_of_week(x)
+    cat(sprintf("[%d cases from %s weeks %s to %s]\n",
+                sum(x$n), type_of_week, head(x$weeks, 1), tail(x$weeks, 1)))
   }
   if (!is.null(group_names(x))) {
     groups.txt <- paste(group_names(x), collapse = ", ")
@@ -21,6 +22,8 @@ print.incidence <- function(x, ...) {
   if (is.integer(x$interval)) {
     cat(sprintf("$interval: %d %s\n",
                 x$interval, ifelse(x$interval < 2, "day", "days")))
+  } else if (grepl("\\d",  x$interval)) {
+    cat(sprintf("$interval: %s\n", x$interval))
   } else {
     cat(sprintf("$interval: 1 %s\n", x$interval))
   }